Fixed bug in inst_size (no longer only returns 1)

+++ src/chunk.rs
@@ -9,7 +9,7 @@ pub enum OpCode {
 }
 
 impl OpCode {
-    pub fn inst_size(self) -> usize {
+    pub fn inst_size(self) -> u8 {
         match self {
             Nop => 1,
             Constant => 3,
